+++ swim/events.go
@@ -244,33 +244,3 @@ func on(t interface{}, f func(e events.Event)) ListenerFunc {
 		}
 	})
 }
-
-// ExecuteThenWaitFor executes a function and then waits for a specific type
-// of event to occur.
-//
-// Often we want to execute some code and then wait for an event be emitted due
-// to the code being executed. However in order to not miss the event we must
-// first register an event handler before we can execute the code:
-// - register listener that signals we can continue on receiving the correct event;
-// - execute the code that will lead to an event being emitted;
-// - wait for the continue signal.
-//
-// This can be quite hard to follow. Ideally we want it to look like
-// - execute the code that will lead to an event being emitted;
-// - and then wait for a specific event.
-//
-// This function helps with making the code read like the latter.
-func ExecuteThenWaitFor(f func(), er events.EventRegistrar, t interface{}) {
-	block := make(chan struct{})
-
-	er.RegisterListener(on(t, func(e events.Event) {
-		select {
-		case <-block: // channel is closed, do nothing
-		default:
-			close(block)
-		}
-	}))
-
-	f()
-	<-block
-}
