+++ swim/handlers_test.go
@@ -93,10 +93,20 @@ func (s *HandlerTestSuite) TestGossipStopHandler() {
 }
 
 func (s *HandlerTestSuite) TestPartitionHealerHandler() {
-	ExecuteThenWaitFor(func() {
-		_, err := s.testNode.node.partitionHealerHandler(s.ctx, &emptyArg{})
-		s.NoError(err, "calling handler should not result in error")
-	}, s.testNode.node, DiscoHealEvent{})
+	done := make(chan struct{})
+	go func() {
+		ExecuteThenWaitFor(func() {
+			_, err := s.testNode.node.discoverProviderHealerHandler(s.ctx, &emptyArg{})
+			s.NoError(err, "calling handler should not result in error")
+		}, s.testNode.node, DiscoHealEvent{})
+		close(done)
+	}()
+
+	select {
+	case <-done:
+	case <-time.After(time.Second):
+		s.Fail("test timed out")
+	}
 }
 
 func (s *HandlerTestSuite) TestToggleGossipHandler() {
