+++ test/mocks/swim_node.go
@@ -1,6 +1,9 @@
 package mocks
 
-import "github.com/uber/ringpop-go/swim"
+import (
+	"github.com/uber/ringpop-go/events"
+	"github.com/uber/ringpop-go/swim"
+)
 import "github.com/stretchr/testify/mock"
 
 type SwimNode struct {
@@ -122,6 +125,6 @@ func (_m *SwimNode) Ready() bool {
 }
 
 // RegisterListener provides a mock function with given fields: l
-func (_m *SwimNode) RegisterListener(l swim.EventListener) {
+func (_m *SwimNode) RegisterListener(l events.EventListener) {
 	_m.Called(l)
 }
